refactor: ManageGroup 그룹 삭제 컴포넌트 분리
