+++ src/components/domain/ManageGroup/index.jsx
@@ -1,1 +1,2 @@
 export { default as ManageMember } from './ManageMember';
+export { default as GroupDelete } from './GroupDelete';
